+++ resources/js/components/index.ts
@@ -7,3 +7,4 @@ export { default as Badge } from './Badge.vue';
 export { default as Button } from './Button.vue';
 export { default as VoteButtons } from './VoteButtons.vue';
 export { default as HelpWantedSign } from './HelpWantedSign.vue';
+export { default as ToolCard } from './ToolCard.vue';